cga_at is two byte registers in and a DI out
It took uint32_t twice and returned uint32_t, while every caller hands it
a position out of a uint8_t field - b->x, h->x, a masked y - and every
caller puts the answer where an offset goes. So it is uint8_t, uint8_t
and uint16_t, which is what the machine holds them in.

224 bytes of .text, for the same reason the row steppers cost 232: the
truncation stops being free once the type says it happens. Checked the
same way - the eight routines that reach video memory through it agree
with the original 10 calls each, and 2,000 frames on level 11, 1,500 on
level 4 and 1,200 through the bonus are identical throughout.

+++ src/game.c
@@ -1655,9 +1655,12 @@ retry:
  *     if (y & 1) di += 0x2000           odd scan lines live in the far half
  *     di += (y >> 1) * 80               `shl ax,4` then `shl ax,2` twice more
  */
-static uint32_t cga_at(uint32_t x, uint32_t y)
+/* A pixel position to the byte that holds it. x and y are the byte registers
+ * the original keeps a position in, and the answer is DI, so all three are
+ * the width they are in the machine. */
+static uint16_t cga_at(uint8_t x, uint8_t y)
 {
-    uint32_t di = x >> 2;
+    uint16_t di = x >> 2;
     if (y & 1)
         di += CGA_PLANE;
     return di + (y >> 1) * CGA_STRIDE;
